+++ src/main/java/net/meander/subtlyd/tags/EntityTypeTagsSD.java
@@ -11,9 +11,9 @@ import net.minecraft.world.entity.EntityType;
 public class EntityTypeTagsSD {
     public static final TagKey<EntityType<?>> CAN_BE_SCARED = create("can_be_scared");
     public static final TagKey<EntityType<?>> SEEKS_SHELTER = create("seeks_shelter");
-    public static final TagKey<EntityType<?>> SEEKS_WARMTH = create("can_seek_warmth"); // TODO rename
-    public static final TagKey<EntityType<?>> SEEKS_SHADE = create("can_seek_shade"); // TODO Rename
-    public static final TagKey<EntityType<?>> PREDATOR = create("can_be_full"); // TODO Rename
+    public static final TagKey<EntityType<?>> SEEKS_WARMTH = create("seek_warmth");
+    public static final TagKey<EntityType<?>> SEEKS_SHADE = create("seek_shade");
+    public static final TagKey<EntityType<?>> PREDATOR = create("predator");
     public static final TagKey<EntityType<?>> NOCTURNAL = create("nocturnal");
     public static final TagKey<EntityType<?>> FEAST_OR_FAMINE_HUNTER = create("feast_or_famine_hunter");
     public static final TagKey<EntityType<?>> SCANSORIAL = create("scansorial");